pyproject.toml
[project]
name = "testing-gemini-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.73.1",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "typer>=0.25.0",
]

[tool.uv]
package = true

[project.scripts]
wtf = "wtf_python.main:app"
